Make half-floats and Eigen vectors work
Turned off warning C4293 for the PreciseHalfCompressor, because it's
moot.

+++ src/math/Math.cpp
@@ -170,7 +170,7 @@ inline int Math::floori(float x)
 
 inline float Math::trunc(float x)
 {
-	return (int)x;
+	return (float)((int)x);
 }
 
 inline int Math::trunci(float x)
